requirements_linux.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854167867/work
certifi==2022.12.7
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046050867/work
channels==4.0.0
channels-redis==4.0.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
click @ file:///tmp/build/80754af9/click_1646056706450/work
colorama==0.4.6
constantly==15.1.0
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1652967019182/work
daphne==4.0.0
Django==4.1.7
django-channels==0.7.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
flit_core==3.8.0
hyperlink==21.0.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
incremental==22.10.0
joblib @ file:///croot/joblib_1666298844297/work
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
msgpack==1.0.4
nltk @ file:///opt/conda/conda-bld/nltk_1645628263994/work
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pymongo==3.12.1
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1663846997386/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytz==2022.7
redis==4.4.1
regex @ file:///tmp/abs_41f5bce5-0a2e-45aa-b231-1fd2fbd57753gfpe6sjm/croots/recipe/regex_1658257178822/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1673863902341/work
requests-oauthlib==1.3.1
service-identity==21.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sqlparse==0.2.4
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
Twisted==22.10.0
txaio==22.2.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
tzdata==2022.7
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1673452138552/work
win-inet-pton==1.1.0
wincertstore==0.2
zope.interface==5.5.2
